fix mypy, add CI

+++ iOpt/method/sol_value.py
@@ -1,12 +1,10 @@
-import numpy as np
-
 from iOpt.trial import FunctionValue
 
 
 class SolutionValue(FunctionValue):
     def __init__(self,
                  calculationsNumber: int = -1,
-                 holderConstantsEstimations: np.double = -1.0
+                 holderConstantsEstimations: float = -1.0
                  ):
         self.calculationsNumber = calculationsNumber
         self.holderConstantsEstimations = holderConstantsEstimations
